Adjust the initial hand positions.

+++ Assets/Scripts/Gameplay/Card.cs
@@ -210,7 +210,7 @@ public class Card : MonoBehaviour
         this.eventsController.MarkCardAsFinishedResolving();
     }
 
-    private IEnumerator AdjustPositionInHand(int index, int cardsInHand)
+    private IEnumerator AdjustPositionInHand(float index, float cardsInHand)
     {
         Vector3 finalPosition;
         switch (this.playerId)
